api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["quan"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.17.6"
peewee = "^3.14.10"
psycopg2-binary = "^2.9.3"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
firebase-admin = "^6.5.0"
python-multipart = "^0.0.9"
bcrypt = "3.2.0"
pyjwt = "^2.9.0"
mangum = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
